Build warning not in Visual Studio's Error List with "Build + IntelliSense" selected. (Visual Studio v17.5.4.)
This project can now be used to demonstrate that the items displayed in the Error List window when "Build + IntelliSense" is selected are not always the union of the items displayed when "Build" is selected and the items displayed when "IntelliSense" is selected. This project might demonstrate a bug in Visual Studio.

Compiling Program.cs generates a CS8602 warning ("Dereference of a possibly null reference.") for line 9. The warning only appears in the Error List when "Build" is selected but not when "Build + IntelliSense" is selected if Program.cs was open in a Visual Studio editor window at the time it was compiled. (Though if Program.cs wasn't open in an editor window when it was compiled then the CS8602 warning does appear when "Build + IntelliSense" is selected. To see this use "Build/Rebuild Solution" if necessary to force recompilation of Program.cs.)

If Program.cs was open in an editor when it was compiled then a CS8632 warning, rather than a CS8602, is shown in the Error List when "Build + IntelliSense" is selected.